Simplify cube physics runtime

+++ engine/helengine.physics3d/runtime/PhysicsSceneFeatureSymbolCatalog3D.cs
@@ -125,7 +125,7 @@ namespace helengine {
                 throw new ArgumentException("Symbol must be provided.", nameof(symbol));
             }
 
-            if ((featureFlags & requiredFeature) == 0) {
+            if (((uint)featureFlags & (uint)requiredFeature) == 0u) {
                 return;
             }
